requirements.txt
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
deprecation==2.1.0
Django==5.1
django-filter==24.3
djangorestframework==3.15.2
et-xmlfile==1.1.0
hightime==0.2.2
idna==3.7
nidaqmx==1.0.0
numpy==2.1.0
openpyxl==3.1.5
packaging==24.1
pandas==2.2.2
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.1
requests==2.32.3
setuptools==73.0.0
simplejson==3.19.3
six==1.16.0
sqlparse==0.5.1
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.2
